Fixed (int)getDouble(...) to getInt(...). Added ExternGeneralizedIO for hydraulic connections.

+++ kernel/mbsim/integrators/lsode_integrator.cc
@@ -155,7 +155,7 @@ namespace MBSim {
     e=element->FirstChildElement(MBSIMINTNS"minimalStepSize");
     setMinimalStepSize(Element::getDouble(e));
     e=element->FirstChildElement(MBSIMINTNS"numberOfMaximalSteps");
-    setmaxSteps(int(Element::getDouble(e)));
+    setmaxSteps(Element::getInt(e));
     setStiff(element->FirstChildElement(MBSIMINTNS"stiffModus"));
       
   }
